Removed .c files without any relevant static content.
