+++ server/src/bpf/csum.h
@@ -20,13 +20,42 @@ Augsburg-Traceroute. If not, see <https://www.gnu.org/licenses/>.
 #ifndef BPF_CSUM_H
 #define BPF_CSUM_H
 
-#include "internal.h"
 #include "ip_generic.h"
 #include <linux/types.h>
+#include <bpf/bpf_endian.h>
 
 // Computes the checksum. See RFC1071 for details.
-INTERNAL __sum16 csum(const void *cursor, __u16 len, __be32 seed);
-INTERNAL __be32 pseudo_header(const iphdr_t *ip, __u16 probe_len,
-                              __u8 protocol);
+static __inline __sum16 csum(const void *cursor, __u16 len, __be32 seed)
+{
+    __be32 sum = seed;
+    const __be16 *pos = cursor;
+
+    while (len > 1) {
+        sum += *(pos++);
+        len -= 2;
+    }
+
+    if (len > 0)
+        sum += *(__u8 *)pos;
+
+    // Fold the recorded carry-outs back into the 16-bit sum.
+    while (sum >> 16)
+        sum = (sum & 0xffff) + (sum >> 16);
+
+    return (__sum16)~sum;
+}
+
+static __inline __be32 pseudo_header(const iphdr_t *ip, __u16 probe_len, __u8 protocol)
+{
+    __be32 pseudo_hdr = bpf_htons(probe_len);
+
+    for (int counter = 0; counter < (sizeof(ipaddr_t) / 2); counter++) {
+        pseudo_hdr += *((__be16 *)&(ip->saddr) + counter);
+        pseudo_hdr += *((__be16 *)&(ip->daddr) + counter);
+    }
+
+    pseudo_hdr += bpf_htons(protocol);
+    return pseudo_hdr;
+}
 
 #endif
