+++ server/src/bpf/cursor.h
@@ -20,19 +20,13 @@ Augsburg-Traceroute. If not, see <https://www.gnu.org/licenses/>.
 #ifndef BPF_CURSOR_H
 #define BPF_CURSOR_H
 
-#include "internal.h"
 #include "ip_generic.h"
+#include <linux/bpf.h>
 
 struct cursor {
     struct __sk_buff *skb;
     void *pos;
 };
-
-INTERNAL long cursor_start(const struct cursor *cursor);
-INTERNAL long cursor_end(const struct cursor *cursor);
-INTERNAL void cursor_reset(struct cursor *cursor);
-INTERNAL void cursor_init(struct cursor *cursor, struct __sk_buff *skb);
-INTERNAL void cursor_clone(const struct cursor *original, struct cursor *clone);
 
 #define PARSE(cursor, hdr)                                                     \
     ({                                                                         \
@@ -46,7 +40,81 @@ INTERNAL void cursor_clone(const struct cursor *original, struct cursor *clone);
         __ret;                                                                 \
     })
 
-// Parses the IP header including any following options.
-INTERNAL int PARSE_IP(struct cursor *cursor, iphdr_t **hdr, __u8 *const proto);
+static __inline long cursor_start(const struct cursor *cursor)
+{
+    return cursor->skb->data;
+}
+
+static __inline long cursor_end(const struct cursor *cursor)
+{
+    return cursor->skb->data_end;
+}
+
+static __inline void cursor_reset(struct cursor *cursor)
+{
+    cursor->pos = (void *)cursor_start(cursor);
+}
+
+static __inline void cursor_init(struct cursor *cursor, struct __sk_buff *skb)
+{
+    cursor->skb = skb;
+    cursor_reset(cursor);
+}
+
+static __inline void cursor_clone(const struct cursor *original, struct cursor *clone)
+{
+    *clone = *original;
+}
+
+static __inline int PARSE_IP(struct cursor *cursor, iphdr_t **hdr, __u8 *const proto)
+{
+    if (PARSE(cursor, hdr) < 0)
+        return -1;
+
+#if defined(TRACEROUTE_V4)
+    long new_pos = (long)(*hdr) + (**hdr).ihl * 4;
+    if (new_pos > cursor_end(cursor))
+        return -1;
+
+    cursor->pos = (void *)new_pos;
+    *proto = (**hdr).protocol;
+    return 0;
+
+#elif defined(TRACEROUTE_V6)
+    struct ipv6_opt_hdr *ext_hdr;
+    __u8 next_hdr = (**hdr).nexthdr;
+
+    for (int i = 0; i < 10; i++) {
+        switch (next_hdr) {
+        case 0:
+        case 43:
+        case 44:
+        case 51:
+        case 50:
+        case 60:
+        case 135:
+        case 139:
+        case 140:
+        case 253:
+        case 254:
+            if (PARSE(cursor, &ext_hdr) < 0)
+                return -1;
+
+            long new_pos = (long)(ext_hdr) + ((ext_hdr->hdrlen + 1) * 8);
+            if (new_pos > cursor_end(cursor))
+                return -1;
+
+            cursor->pos = (void *)new_pos;
+            next_hdr = ext_hdr->nexthdr;
+            continue;
+        default:
+            break;
+        }
+    }
+
+    *proto = next_hdr;
+    return 0;
+#endif
+}
 
 #endif
