+++ server/src/bpf/resize.h
@@ -20,14 +20,43 @@ Augsburg-Traceroute. If not, see <https://www.gnu.org/licenses/>.
 #ifndef BPF_RESIZE_H
 #define BPF_RESIZE_H
 
-#include "internal.h"
-#include "ip_generic.h"
-#include <linux/types.h>
+#include "cursor.h"
+#include <linux/bpf.h>
+#include <linux/if_ether.h>
+#include <linux/ip.h>
+#include <bpf/bpf_endian.h>
+#include <bpf/bpf_helpers.h>
 
 struct cursor;
 struct ethhdr;
 
-INTERNAL int resize_l3hdr(struct cursor *cursor, __u16 probe_len,
-                          struct ethhdr **eth, iphdr_t **ip);
+/*
+ * Resizes the packet to be able to hold the specified length on top of the IP
+ * header, all options are truncated.
+ */
+static __inline int resize_l3hdr(struct cursor *cursor, __u16 probe_len,
+                          struct ethhdr **eth, iphdr_t **ip)
+{
+    int ret = bpf_skb_change_tail(cursor->skb,
+                                  sizeof(**eth) + sizeof(**ip) + probe_len, 0);
+
+    if (ret < 0)
+        return -1;
+
+    cursor_reset(cursor);
+    if (PARSE(cursor, eth) < 0)
+        return -1;
+    if (PARSE(cursor, ip) < 0)
+        return -1;
+
+#if defined(TRACEROUTE_V4)
+    (**ip).ihl = 5;
+    (**ip).tot_len = bpf_htons(cursor->skb->len - sizeof(**eth));
+#elif defined(TRACEROUTE_V6)
+    (**ip).payload_len = bpf_htons(probe_len);
+#endif
+
+    return 0;
+}
 
 #endif
